@@ -116,90 +118,89 @@
     ]
   },
   "devDependencies": {
-    "babel-core": "^6.24.1",
-    "babel-eslint": "^7.2.3",
-    "babel-jest": "^20.0.3",
-    "babel-loader": "^7.1.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.1",
+    "babel-jest": "^22.1.0",
+    "babel-loader": "^7.1.2",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-dev-expression": "^0.2.1",
-    "babel-plugin-dynamic-import-webpack": "^1.0.1",
-    "babel-plugin-flow-runtime": "^0.11.1",
+    "babel-plugin-flow-runtime": "^0.15.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-es2015-classes": "^6.24.1",
-    "babel-preset-env": "^1.5.1",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-react-optimize": "^1.0.1",
     "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.24.1",
-    "babili-webpack-plugin": "^0.1.2",
-    "chalk": "^2.0.1",
-    "concurrently": "^3.5.0",
-    "cross-env": "^5.0.0",
-    "cross-spawn": "^5.1.0",
-    "css-loader": "^0.28.3",
-    "electron": "^1.6.10",
-    "electron-builder": "^19.8.0",
-    "electron-devtools-installer": "^2.2.0",
-    "enzyme": "^2.9.1",
-    "enzyme-to-json": "^1.5.1",
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^15.0.1",
-    "eslint-formatter-pretty": "^1.1.0",
-    "eslint-import-resolver-webpack": "^0.8.3",
-    "eslint-plugin-compat": "^1.0.4",
-    "eslint-plugin-flowtype": "^2.33.0",
-    "eslint-plugin-flowtype-errors": "^3.3.0",
-    "eslint-plugin-import": "^2.6.0",
-    "eslint-plugin-jest": "^20.0.3",
-    "eslint-plugin-jsx-a11y": "5.0.3",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-react": "^7.1.0",
-    "express": "^4.15.3",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "fbjs-scripts": "^0.8.0",
-    "file-loader": "^0.11.1",
-    "flow-bin": "^0.48.0",
-    "flow-runtime": "^0.13.0",
-    "flow-typed": "^2.1.2",
-    "html-webpack-plugin": "^2.29.0",
+    "babel-register": "^6.26.0",
+    "chalk": "^2.3.0",
+    "concurrently": "^3.5.1",
+    "cross-env": "^5.1.3",
+    "cross-spawn": "^6.0.4",
+    "css-loader": "^0.28.9",
+    "detect-port": "^1.2.2",
+    "electron": "^1.7.11",
+    "electron-builder": "^19.55.3",
+    "electron-devtools-installer": "^2.2.3",
+    "electron-rebuild": "^1.7.3",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "enzyme-to-json": "^3.3.1",
+    "eslint": "^4.16.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-formatter-pretty": "^1.3.0",
+    "eslint-import-resolver-webpack": "^0.8.4",
+    "eslint-plugin-compat": "^2.2.0",
+    "eslint-plugin-flowtype": "^2.42.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-jest": "^21.7.0",
+    "eslint-plugin-jsx-a11y": "6.0.3",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-react": "^7.6.1",
+    "express": "^4.16.2",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "fbjs-scripts": "^0.8.1",
+    "file-loader": "^1.1.6",
+    "flow-bin": "^0.64.0",
+    "flow-runtime": "^0.16.0",
+    "flow-typed": "^2.3.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^20.0.4",
-    "jsdom": "^11.0.0",
+    "jest": "^22.1.4",
     "minimist": "^1.2.0",
-    "node-sass": "^4.5.3",
-    "react-addons-test-utils": "^15.6.0",
-    "react-test-renderer": "^15.6.1",
+    "node-sass": "^4.7.2",
+    "npm-logical-tree": "^1.2.1",
+    "react-test-renderer": "^16.2.0",
     "redux-logger": "^3.0.6",
-    "rimraf": "^2.6.1",
+    "rimraf": "^2.6.2",
     "sass-loader": "^6.0.6",
-    "sinon": "^2.3.5",
-    "spectron": "^3.7.0",
-    "style-loader": "^0.18.1",
+    "sinon": "^4.2.2",
+    "spectron": "^3.8.0",
+    "style-loader": "^0.20.1",
     "stylefmt": "^6.0.0",
-    "stylelint": "^7.12.0",
-    "stylelint-config-standard": "^16.0.0",
-    "url-loader": "^0.5.8",
-    "webpack": "^3.0.0",
-    "webpack-bundle-analyzer": "^2.8.2",
-    "webpack-dev-server": "^2.5.0",
-    "webpack-merge": "^4.1.0"
+    "stylelint": "^8.4.0",
+    "stylelint-config-standard": "^18.0.0",
+    "uglifyjs-webpack-plugin": "1.1.8",
+    "url-loader": "^0.6.2",
+    "webpack": "^3.10.0",
+    "webpack-bundle-analyzer": "^2.9.2",
+    "webpack-dev-server": "^2.11.1",
+    "webpack-merge": "^4.1.1"
   },
   "dependencies": {
     "devtron": "^1.4.0",
-    "electron-debug": "^1.2.0",
+    "electron-debug": "^1.5.0",
     "font-awesome": "^4.7.0",
-    "history": "^4.6.3",
-    "react": "^16.1.0",
-    "react-dom": "^15.6.1",
-    "react-hot-loader": "3.0.0-beta.6",
-    "react-redux": "^5.0.5",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
+    "history": "^4.7.2",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-hot-loader": "^4.0.0-beta.13",
+    "react-redux": "^5.0.6",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "react-router-redux": "^5.0.0-alpha.6",
-    "redux": "^3.7.1",
+    "redux": "^3.7.2",
     "redux-thunk": "^2.2.0",
-    "source-map-support": "^0.4.15"
+    "source-map-support": "^0.5.3"
   },
   "devEngines": {
     "node": ">=7.x",
