@@ -3,7 +3,11 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import styles from './Home.css';
 
-export default class Home extends Component {
+type Props = {};
+
+export default class Home extends Component<Props> {
+  props: Props;
+
   render() {
     return (
       <div>
