@@ -2,7 +2,11 @@
 import React, { Component } from 'react';
 import Home from '../components/Home';
 
-export default class HomePage extends Component {
+type Props = {};
+
+export default class HomePage extends Component<Props> {
+  props: Props;
+
   render() {
     return (
       <Home />
