@@ -1,20 +1,22 @@
 // @flow
-import React from 'react';
+import React, { Component } from 'react';
 import { Provider } from 'react-redux';
 import { ConnectedRouter } from 'react-router-redux';
 import Routes from '../routes';
 
-type RootType = {
+type Props = {
   store: {},
   history: {}
 };
 
-export default function Root({ store, history }: RootType) {
-  return (
-    <Provider store={store}>
-      <ConnectedRouter history={history}>
-        <Routes />
-      </ConnectedRouter>
-    </Provider>
-  );
+export default class Root extends Component<Props> {
+  render() {
+    return (
+      <Provider store={this.props.store}>
+        <ConnectedRouter history={this.props.history}>
+          <Routes />
+        </ConnectedRouter>
+      </Provider>
+    );
+  }
 }
