@@ -22,7 +22,11 @@ const configureStore = (initialState?: counterStateType) => {
     level: 'info',
     collapsed: true
   });
-  middleware.push(logger);
+
+  // Skip redux logs in console during the tests
+  if (process.env.NODE_ENV !== 'test') {
+    middleware.push(logger);
+  }
 
   // Router Middleware
   const router = routerMiddleware(history);
@@ -52,8 +56,7 @@ const configureStore = (initialState?: counterStateType) => {
 
   if (module.hot) {
     module.hot.accept('../reducers', () =>
-      store.replaceReducer(require('../reducers')) // eslint-disable-line global-require
-    );
+      store.replaceReducer(require('../reducers'))); // eslint-disable-line global-require
   }
 
   return store;
