@@ -9,15 +9,11 @@ function CheckBuildsExist() {
   const rendererPath = path.join(__dirname, '..', '..', 'app', 'dist', 'renderer.prod.js');
 
   if (!fs.existsSync(mainPath)) {
-    throw new Error(chalk.whiteBright.bgRed.bold(
-      'The main process is not built yet. Build it by running "npm run build-main"'
-    ));
+    throw new Error(chalk.whiteBright.bgRed.bold('The main process is not built yet. Build it by running "npm run build-main"'));
   }
 
   if (!fs.existsSync(rendererPath)) {
-    throw new Error(chalk.whiteBright.bgRed.bold(
-      'The renderer process is not built yet. Build it by running "npm run build-renderer"'
-    ));
+    throw new Error(chalk.whiteBright.bgRed.bold('The renderer process is not built yet. Build it by running "npm run build-renderer"'));
   }
 }
 
