@@ -7,9 +7,7 @@ export default function CheckNodeEnv(expectedEnv: string) {
   }
 
   if (process.env.NODE_ENV !== expectedEnv) {
-    console.log(chalk.whiteBright.bgRed.bold(
-      `"process.env.NODE_ENV" must be "${expectedEnv}" to use this webpack config`
-    ));
+    console.log(chalk.whiteBright.bgRed.bold(`"process.env.NODE_ENV" must be "${expectedEnv}" to use this webpack config`));
     process.exit(2);
   }
 }
